requirements.txt
streamlit>=1.33
audio-recorder-streamlit>=0.0.8
soundfile>=0.12
numpy>=1.24
python-dotenv>=1.0

# Transcription backends (default: faster-whisper)
faster-whisper>=1.0
openai>=1.40.0
whisper @ git+https://github.com/openai/whisper.git#egg=whisper

# Snowflake (Search via Snowpark/Core + LLM via SQL COMPLETE)
snowflake-connector-python>=3.10
snowflake-snowpark-python>=1.19
snowflake-core>=0.10
